requirements.txt
bcrypt==4.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2==2.9.5
python-dotenv==0.21.1
Werkzeug==2.2.3
